Layout items in header
Setting the header to display: flex; and hiding the .mobile
elements with display: none; caused flexbox to act weird,
resulting in extra spacing between the right edge and the social
media icons. An extra div containing only the items which should
be on the top row was added and given display: flex; instead,
resolving this isssue.